bpf: assembler-side changes.
The old DIF assembler went digging through the instruction stream to
identify jumps and items requiring relocation, and then fixed them up.

Use the change to BPF to make this less pointlessly arcane and tightly-
coupled: instead of recording things into the instruction stream and
then decoding them later (relying on the details of the format we
emitted from dt_cg), simply record what is needed into new fields in the
irlist, di_subr_op (the subr called when a DIF function or subr is
translated) and di_extern_int (the integer needed to relocate when
di_extern is set).

I am even less confident than usual in the details of the relocation
machinery and the value relocated: it needs more testing than most of
this will.

In the process, augment dt_cg_call() to allow it to pass back a pointer
to the call instruction generated (we can't just use the last because it
may be followed by a bunch of unspills) and adjust all callers.

Where dynamic translator machinery gets in the way, we simply delete
it, since it is only half-implemented and there is no point preserving
the generation of opcodes that nothing knows how to execute.