bpf: convert DIF programs in the compiler into BPF.
The compiler contains a few tiny programs that generate constants,
mostly for use as subr arguments and the like. We arbitrarily assume
that a suitable program is one that does a load of r0, then an exit:
it may well be that we don't want the exit in this case, if (say) we end
up using bpf-to-bpf calls or something for this.