dlibs: suppress spurious whitespace around define_for_kernel()
define_for_kernel() merely defines a macro: it is not supposed to
emit anything. However, the macro contains a great deal of indentation
which is translated unchanged into the output, leading to several
hundred columns of tabs and spaces on every invocation.

Suppress this with a diversion. (A single space and linefeed remains,
but this is still much better than it was.)

Add a few missing m4_dnls around some of the uses, too, which were
leading to double-linefeeds in the output.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>