dlibs: add expand_for_kernel() and use it
We have a way of defining macros with values depending on the kernel
version, but no way to decide to simply emit ordinary text depending on
the kernel version.

Now we have suppressed most whitespace from define_for_kernel, this is
easy: just use it with a macro, then immediately expand that macro.

Use this to make PIDTYPE_TGID vanish entirely in kernel 4.18 and below
(where it does not exist.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>