dt_printf: make threadsafe
In an upcoming commit we will start invoking dt_printf() from multiple
threads.  While stdio handles locking in some cases, the caller of
dtrace(1) decides (via the undocumented -B argument) whether to use
buffered output, and if buffered output is in use we must protect the
buffering from concurrent modification.

While we're at it, use the same lock to protect the buffer used for the
sprintf() action, just in case: though it seems unlikely this is truly
needed, the consequences if it is and are not present would be severe.
(-B is rarely-used enough that there is no point having two locks for
this: name the lock after sprintf(), since it's actually documented and
used by real people.)

Orabug: 28953618
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>