cc: use pthread_sigmask()
This code was using sigprocmask() to block SIGCHLD around preprocessor
invocation.  This was always a bad idea even back in the Solaris days,
because the process-monitoring threads were always, well, threads, and
sigprocmask() is undefined in the presence of threads (even if on nearly
all platforms it usually works in practice).

Use pthread_sigmask() instead.

Orabug: 28953618
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>