Add pcap(skb, proto) action to support libpcap-based packet capture
Currently this action supports two modes of operation.  When a file has been
freopen()ed, the pcap() action will pcap_dump() to the specified file.
For example:

ip:::send {
	freopen("/tmp/cap.%s", probename);
	pcap((void *)arg0, PCAP_IP);
	freopen("");
}

The final freopen("") ensures that other DTrace output is not dumped to the
file also; using "-q" avoids the need for this.  Note also that since
freopen() is a destructive action, the above script would need to be
invoked with the "-w" option.

If a file is not freopen()ed, the capture data will be passed to an
invocation of tshark via a named pipe and displayed with a leading
prefix of 'pcap '; if tshark is not available or we fail to execute it
for some reason the output will be displayed in a tracemem-like format.
Example usage:

dtrace -n 'ip:::send { pcap((void *)arg0, PCAP_IP); }'

One minor caveat for libdtrace consumers: if tshark is in use, and your
application intercepts SIGCHLD, it should expect SIGCHLDs from tshark
termination.  (We cannot mask them because tshark is a long-running
process which might die at any time, since we do not control it and it
will die if it's hit by a signal, if it runs out of memory, etc...)

[nca: improve piping to tshark; pipe output back via an output thread]

Orabug: 28953618
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>