pcap: improve security of tshark invocation
We are running tshark as root, which is horribly dangerous: indeed,
tshark even warns us about this but it does so on stderr, which we've
thrown away.  tshark does try to relinquish privileges, but only by
setting its uid and gid back to the real ones, which doesn't help us
because dtrace isn't suid root.

Unfortunately, dropping our permissions is a bit tricky.  We can't just
drop the lot and sanitize the entire environment because tshark tries to
open files out of $HOME, and fails if it can't read it, so we must point
$HOME somewhere readable (/run/initramfs for now, which must always be
present on a running system and is world-readable).  We also have to set
the group to 'wireshark', at least on OL7 where dumpcap is setgid to
that gid and not executable by other.  We set the uid to -3 because -1
is 'nobody' and -2 is often used by people trying to lower privs without
stamping on 'nobody' and on some systems is itself 'nobody'.

Also improve synchronization between dtrace and tshark, so we wait for
tshark to start before feeding packets to it and so we can diagnose
tshark startup failures.  This is a little tricky because we cannot use
stdout for this (stdout output is routine) but it would seem that we
also cannot use stderr, since we don't want to hang on to a stderr pipe
forever.  But what we can do is set up a stderr pipe and mark it
O_CLOEXEC, os that as soon as the exec succeeds the pipe goes away and
we can consider the tshark started.  This also lets us read the error
out of stderr and feed it to dt_dprintf().  The only thing we still
can't do is get fatal errors from tshark if it starts and then fails to
function -- but that doesn't matter, because the printing thread already
detects tshark failure (although it cannot find out why it failed), and
can we can arrange for the printing thread to flip the same tracemem
switch so that dtrace will immediately switch to tracemem()-like output.

We probably want to make the homedir location and possibly the
unprivileged uid and wireshark gid either customizable options or a
compile-time option (to integrate with a new distro-specific tsharking
uid).

The only downside of this patch other than all the distro-specific
customizable stuff is that tshark drops a tempfile in /tmp while it
runs, and that tempfile will have a very strange uid.  This is
unavoidable, really, and actually serves as a good test that things are
working.

We add a new test to verify that tracemem() fallback on fork failure
works.

Orabug: 28953618
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>