pcap: mask SIGINT in tshark child
Alan's original submission prevented SIGINT submission to tshark by
setsid().  I took this out because it seemed unnecessary, but after
recent changes to detect tshark dying and switch to tracemem output
dynamically it has become clear that we are subject to a race when
DTrace is interrupted, where tshark can die before the DTrace probes
have emptied, and then it sprays a pile of tracemem output at the
screen.

So mask SIGINT (the only signal the user is likely to interrupt us with)
before invoking tshark.

Orabug: 28953618
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>