libproc: fix return-without-value errors in earlier rd_delete() split fix
Commit ba18f771bde29e10a691874f8f68e43126a4d108 split up rd_delete()
into rd_split() and rd_free() and rejigged the rtld_db public entry
points to detect deleted-but-not-freed rd_agents...

... or that was what was meant, but unfortunately these entry points
often did bare returns when the functions they were part of return
values.  This is highly suboptimal (and generated entirely accurate
compiler warnings which I persistently failed to see for months).

Fix by returning errors in all these cases, though the value to return
is debatable.  Returning an error seems like a more appropriate return
value than returning success, and it is more likely that callers will do
appropriate things in that case: we do return success in similar-seeming
on-entry cases in rd_event_enable(), but that case is checking if a live
process has monitoring already enabled, in which case the call is a NOP.
It's still an error to try to enable monitoring on a dead process...

(Audited callers: they do all do more sensible things for this case on
error returns than on success returns.)

Orabug: 28990529
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>