libproc: fix the rtld_db lifetime properly, without leaking
Commit ba18f771bde29e10a691874f8f68e43126a4d108 (and the subsequent
fixup commit 0b6f0142ffd59f2b74169d0c01a67e3c2871e4db) tried to fix the
rtld_db lifetime so that rtld_db wouldn't vanish between calls, but
would get neatly released when Prelease() was called, and freed when
Pfree() was called.

Unfortunately the variable it chose to use as a flag that a release had
happened (so calls to the public API should do nothing) was badly
chosen: rd->P->rap is the pointer to the rd_agent inside the libproc
handle, and setting it to NULL as an indication that a release has
happened (but a free has not) necessarily prevents Pfree_internal() from
freeing it at all (since P->rap is the only pointer to a rd_agent it can
see). So we end up always leaking the rd_agent.

Instead, use a dedicated release flag to do the same thing.  (r_brk_addr
*almost* suffices, but unfortunately a released rd_agent with a
still-live process can cause a zero r_brk_addr to go nonzero again.)

Orabug: 28990529
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>