libproc, ptrace: do not leak a state word on error or get out of sync
When Ptrace() is called, it calls the ptrace lock hook with 1 if the
ptrace() count is zero, increments the ptrace() count, and pushes the
previous process state onto the state stack.  When Puntrace() is called,
it reverses all of this.

But when Ptrace() fails with an error, it fails to pop the state, though
it does decrement the ptrace count.  This leaks one state stack entry
forever. (It also leads to the state and ptrace count getting out of
sync, but since this error path is never entered unless things have gone
very wrong, the state of the process is probably irrelevant anyway
because dtrace is just about to terminate with a horrible error
message.)

(Note that not *all* pushes need to be accompanied by pops, because we
might exec() in the middle, in which case all the state pops will have
been done by the time Ptrace() finishes.  We can detect this case by
checking the ptrace_count at decrement time, as Puntrace() does.)

Orabug: 29036502
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>