pid: prohibit self-grabs with a slightly nicer message
If you try to put a pid probe on the dtrace consumer itself, you get the
message

dtrace: failed to compile script /dev/stdin: line 1: process 7921 is not a dynamically-linked executable

(with obvious pid substitution). This is obviously ridiculous: what's
actually happening is that we are refusing to grab our own PID for
symbol resolution because that involves briefly stopping and then
restarting ourselves, and it so happens that the first such lookup is
scoped inside ld.so.  Add a message that makes things clearer.

Orabug: 29036502
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>