tests: pid provider tests should not rely on awk triggers
pid provider tests rely on certain symbolic information,
which is not present with stripped executables unless built
with --export-dynamic.

A few pid provider tests used awk as a trigger, even though
this binary is stripped.  Until recently, it was at least built
with --export-dynamic.  That is no longer true with gawk 4.2.1.

Use other triggers for pid provider tests to avoid this problem.

Orabug: 29056323
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>