tests: more robust mechanism for checking ld.so tracing
The dtrace utility supports a "-x evaltime=$stage" option to control
when to start tracing a new process.  In particular, for dynamically
linked binaries, should tracing begin before or after the dynamic linker
runs?  tst.evaltime.sh has assumed that ld.so tracing is indicated if
and only if syscall::brk:entry fires.  This brittle assumption can
break, and it does break with newer glibc (somwhere between 2.17 and
2.27), where other activity, apparently during constructor execution,
also generates brk() calls.  This check also produces the ridiculous
conclusion that static binaries are using the dynamic linker.

"-x evaltime" can also be used to control whether tracing starts
before or after constructors are executed.  tst.evaltime.sh checks this
by having a constructor in the trigger process write a message and
then counting how many times syscall::write:entry fires.  While we
have not seen any shortcomings of this approach, we can improve on it.

Therefore:

- Check for ld.so by tracing on
  'syscall:::entry /pid == $target/ {ustack()}'.
  Look for "_dl_start" on the reported stacks.

- Check more specifically for printed messages by having
  syscall::write:entry also report the length of the message.
  (It would be nice to have the probe report the contents of the
  message instead, but that is currently difficult on ARM, where
  copyinstr does not yet work.)

- Clean up the comments regarding the functionality and the documentation.
  Specifically, eliminate the claim that the functionality is not documented.

The 32-bit version of this test has other issues.  Notably, the underlying
functionality is not working, and multiple bitnesses of a test can be
implemented more elegantly than with cut and paste.  Those issues are
outside the scope of this commit.

Orabug: 29610875
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>