Prevent loss of notification enqueues on process destruction
When a dt_proc_t is destroyed, we remove it from every datastructure it
appears in, to avoid dereferencing freed memory.  This includes removing
outstanding notifications for it on the dt_proc_notify_t list.  For
destruction that happens at dtrace_close() time, this is fine, since the
entire notification list and everything that might query it is also
gone: but short-term process grabs done by things like ustack() can
cause destruction at arbitrary times, and if we destroy a process with
an outstanding notification, that notification is no longer sent.
Normally short-term grabs do not have notification enabled, but if we
were doing a short-term grab of a process that we were also doing a
long-term grab of for -c purposes, it *does* have notifications enabled,
and a death notification can easily be missed, leading to dtrace -c not
terminating even though the process it was monitoring has terminated.

The fix is simple: rather than destroying the entire notification list
entry when a process is destroyed, merely NULL out its dprn_dpr pointer
to the dt_proc: this has minimal negative effects because we don't pass
the dpr to the notifier function: we only need it to lock the dpr around
the notification send, which is definitely unnecessary if the dpr has
been destroyed.  We do want to know the pid, so store that in the
dt_proc_notify itself: retain existing notifier behaviour (where some
negative PID means 'dead process') by negating the pid in the notifier
call (previously we just unconditionally passed in -1, which was a fat
lot of use because the notified code then couldn't tell which process
had died!).

This turns out to be embarrassingly easy to replicate: add a new test to
cover it.

Orabug: 29195206
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>