global: de-bcopy()/bzero()/bcmp()
Erase all three of these deprecated identifiers from DTrace, replacing
them with memcpy(), memset() and memcmp().  (It happens that there were
no uses of bcmp(), but if there had been we'd have caught them.)

99% done via Coccinelle, with a post-facto sedding of the script to
prevent it eliminating the space in 'sizeof (...)' and a few post-facto
reversals of its removal of big comments on the assumption that they
were related to the #include of strings.h.

(We put a #include of string.h into dtrace_impl.h to avoid having to do
it piecemeal all over the tree. The strings.h includes were all over the
place, often in files that didn't need them at all. This is neater.)

Coccinelle semantic patch applied via via

spatch --cocci-file maint/de-bcopy.cocci --prevent-kr --patch . \
	-I include -I uts/common -I include/i386 -I libproc/i386 \
	-I libdtrace/i386 **/*.c | sed 's,sizeof(,sizeof (,g' | less

followed by doing the same over arm64 (to only files in arm64 subdirs,
which is not many), and unification of the patches (arm64 had nothing
new so that was easy) and a little light fixing up of indentation of
continuation lines, which Coccinelle insists on doing in aligned-param-
list Linux kernel style.

valgrind shows no new problems.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>