tests: don't assume we have IPv6 facilities; XFAIL if no IPv6
There are four tests which have been operating on the assumption the system
has IPv6 operating on the system under test.  Where there are some
cases this assumption is not correct, this commit adds checks in the four
test scripts and returns a value of 67 -- indicating XFAIL for the test --
if /sbin/ip route does not find ::1.

Signed-off-by: David  Mc Lean <david.mclean@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>