libproc: fix file_dyn_base when not executing out of the first segment
When determining correspondences between addresses in ELF files and
addresses in memory, one must adjust them to allow for the address the
dynamic linker loaded them at (the 'load bias').  Normally we extract
this from the dynamic linker directly, but when that can't be done, such
as when getting the address of the rtld_global symbol in the dynamic
linker itself, we have to figure it out ourselves.

Unfortunately the method used was less than ideal.  We were iterating
through all the phdrs to find the one with the lowest vaddr (which is
the first segment), then subtracting that vaddr from the start address
of the mapping code is being executed out of.  This is clearly wrong:
you can't subtract the vaddr of one mapping from the start address of a
possibly different mapping and hope to get meaningful results. (We are
also not allowing for segment alignment != PAGE_SIZE, but on Linux this
is rarely a problem).  When the first PT_LOAD mapping is not the one
code is being executed out of, we end up with a load bias that's off by
the difference between the addresses of the two mappings in question.
This rarely has good consequences.

So keep track of the first mapping when we see it in /proc/$pid/maps, in
the new prmap_file.first_segment member.  This gives us access to the
load address of the first segment, not just its vaddr in the ELF file,
so it becomes easy to figure out the difference between the first
mapping's load address and its vaddr.  It also gets easier to find that
vaddr: rather than doing a possibly-lengthy traversal of all the phdrs
just in case there's one later on with a lower vaddr, we only need to
iterate through them until we find the first PT_LOAD one, which
necessarily always corresponds to the first_segment above.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Orabug: 29589029