tests, pldd: filter out the vDSO on later glibc versions
pldd filters out the vDSO by looking for a loadable object with no
search path and no name.  This is not enough for later glibc versions,
which set the name: so also filter out loadable objects with no
search path that happen to be the second one in the list.  (Even this is
more rigorous than what libproc actually does, which is just to say
"if it's the second one in the list, it's the vDSO". This is dangerously
loose, and will be fixed in the next commit.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Orabug: 29589029