libproc: shared objects with a search path cannot be the vDSO
Simply declaring that the vDSO is the second item in the search path is
dangerous: while it is always second if present (a restriction enforced
by the kernel and dpeended on by glibc), the vDSO might not even be
enabled, and then we skip a completely random shared library.

So only skip the vDSO if this looks likely to be the vDSO.  Normal
shared objects always have a search path with length > 1, so we can
use a shorter search path as a more reliable heuristic than the purely
positional one we're using now.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>