libproc: further armouring against invalid DL_NNS values
We know that you can't have no namespaces, but namespaces in glibc are a
small static resource, and we know we know its maximum (because if that
maximum changes, it changes the offsets of stuff in rtld_global that we
need so we *have* to adapt).

So export the maximum number of namespaces to libproc and check it: if
there are more namespaces than that, we know we aren't looking at the
namespace count, or glibc has bumped its DL_NNS value and we must
adapt.  (This is a bitness-independent value.)

Suggested-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Orabug: 29589029