test: funcs/tst.progenyof.d logic needs fixing
The test claims
        progenyof($ppid + 1);  /* this will always be false */
But this claim is false since $ppid+1 could be $pid, and a process
is progeny of itself.  Indeed, when dtrace is launched from "timeout",
as it is in the test suite, we expect $pid to be $ppid+1.

Change the test to use other impossible "progenyof()" relationships,
such as progenyof($pid + 1) or progenyof($trigger).

Also, change the test to check progenyof() for nonzero (rather than
positive) values to comply with the documentation and comments.

[Orabug: 29705902]
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>