Handle empty clauses.
Specifying a probe with an empty clause caused a SEGV because the code
generator tried to free a register associated with the clause.  Since the
clause is empty, no register is ever allocated and the register index that
is being free'd is -1, which is not a valid register.

The solution is trivial: when the clause is empty, there is no need to
invoke the code generator.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>