Remove integer table support.
DTrace used to store integer constants in an integer table, to allow long
integers to be encoded in instructions using less bits.  In BPF we can
assign values directly using a MOV instruction (up to 32-bit) or a LDDW
instruction (64-bit).  There is no need for the integer table anymore.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>