Remove the opening of the DTrace device node.
Since DTrace based on eBPF no longer requires communication with the
kernel by means of the DTrace device node that was created by the DTrace
core kernel module (no longer used), we can do away with the code that
handles it.  Likewise, we no longer use the fasttrap module and its
device node.

THis patch also removed the dt_fdlist struct type that is no longer used
by any of the code.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>