Remove drti.o and related support code.
Since DTrace will no longer have a core kernel module, it will also not be
providing a helper interface to load DOF for an executable.  This patch
therefore also removes the drti.o object that was used to load the embedded
DOF into the kernel by means of the helper interface.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>