Import former kernel DTrace header files.
Future versions of DTrace (which this is the root for) will no longer
comprise a kernel and userspace portion as was the case in the previous
implementation.  As a result, we need to move the DTrace headers into
the userspace source tree because the kernel headers will no longer
provide them.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>