Adjust code to structure changes in dtrace_probedesc_t.
The change from statically sized character arrays for probe name elements
to const char pointers has some far reaching impact on the userspace code.
This patch adjusts code in various areas (incl. libproc) to account for
this.

These strings can originate in various places (where they are allocated),
and are free'd once we are done with them.  The places where we free them
are indicated with comments and a cast to char * to keep the compiler
quiet.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>