Add generic hashtable implementation.
This patch adds a generic hashtable implementation that is primarily used
for hashing probes based on provider, module, function, or probe name,
and on the fullyqualified name (provider:module:function:name).

The hashtable resizes automatically as the distribution of entries over
the hash slots and buckets per slot gets more tight.  Users of this
hashtable implementation must provide entry-specific functions to calculate
the hash value of an entry, compare two entries, add an entry to a list of
entries, and remove an entry from a list of entries.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>