Split probe implementation from provider implementation.
The implementation of probe-specific functionality creating new probes,
probe lookup, ...) was included in the provider implementation in the
dt_provider.c source code file.  Since all provider and probe handling
is not being done at the userspace level, it makes sense to split them
up.  This improves code clarity and it makes it easier to know where to
look for the implementation details of provider and probe handling.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>