Emit disassembly for predicate even if there are no actions.
Previously, the disassembler would not produce any output if there were
no actions associated with it, even if there was a predicate.  An ECB
with predicate but no actions can be used to trace the probe firing
event (without extra data) under conditions set in the predicate.

This patch ensures that disassembly is produced for any and all code in
the ECB, predicate and/or actions.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>