Suppress parse tree (pass 3) output when parsing failed badly.
Sometimes, parsing of the input fails so badly that there is no result
whatsoever.  In that case, pass 1 and 2 parse trees are skipped, yet the
pass 3 parse tree output (when -xtree=4 is supplied) was still being
generated (or at least an attempt was made).  WHen parsing failed
completely, there is no tree to dump and we ended up with a NULL pointer
dereference.

This patch ensures that dtrace does not try to dump a parse tree when
the root of the tree is NULL.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>