Provide better error message for incomplete clauses.
In some cases (e.g. BEGIN/pid==1234/), an incomplete clauses would
result in a 'syntax error near "' message.  The problem is that it
was trying to print the character where the syntax error occured,
which in this case is a 0 byte.  The message now correctly reports
that the syntax error is near the end of the input.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>