dtrace: handle .init.scratch section in /proc/kallmodsyms
The introduction of the .init.scratch section in Linux 5.2-rc6 via e1bfa87
("x86/mm: Create a workarea in the kernel for SME early encryption") led to
the consistent failure of
- test/unittest/aggs/tst.aggmod_full2.sh
- test/unittest/consumer/tst.merge_ranges_bug25767469.c
- test/unittest/consumer/tst.symbols.c
This init section, related to the kernel, appears in /proc/kall[mod]syms
after the end of kernel addresses, causing problems for DTrace's ways of
reading /proc/kallmodsyms.

Until the kernel is fixed not to write this init section to
/proc/kall[mod]syms, add fixes to DTrace userspace code wherever, including
in tests, it reads /proc/kallmodsyms.

[Orabug: 30149066]
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>