Encode built-in BPF functions by id and emit names in disassembler.
Since various operations are being implemented as BPF functions, we need
to store the function id in the call instruction (to be patched later
with the actual function instruction offset.  Disasembler output will
display the proper function name.

This patch also provides a more useful comment for instructions that
access thread-local and clause-local variables, indicating the scope of
the variable rather than just the name.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>