dtrace: correct snprintf maximum lengths in libdtrace/dt_printf.c
dt_printf_format() and dtrace_printf_format() had improperly specified
values for the maximum size argument of snprintf.

dt_printf_format() has two instances of snprintf where the destination
pointer was not at the beginning of the buffer space, while the maximum
size argument specified the total size of the buffer rather than the
remaining length.  This has not been a practical problem because the
output string has always been much shorter than the remaining length of the
buffer.

Tests that failed without the correction of this commit are:
    test/unittest/aggs/tst.keysort.d
    test/unittest/multiaggs/tst.many.d
    test/unittest/multiaggs/tst.sort.d
    test/unittest/multiaggs/tst.sortpos.d
    test/unittest/multiaggs/tst.tuplecompat.d
    test/unittest/multiaggs/tst.zero.d
    test/unittest/multiaggs/tst.zero2.d
    test/unittest/printa/tst.dynwidth.d
    test/unittest/printa/tst.stack.d
    test/unittest/printf/tst.flags.d
    test/unittest/printf/tst.widths.d
    test/unittest/printf/tst.widths1.d
    test/unittest/printf/tst.wp.d
    test/unittest/rates/tst.aggrate.d
    test/unittest/rates/tst.switchrate.d
    test/demo/intro/trussrw.d
    test/demo/struct/rwinfo.d

The error seen while running on Ubuntu 19.04 was:
"*** buffer overflow detected ***: dtrace terminated"

In each case a core dump file was generated.

Similarly, the maximum length argument was seen to be incorrect for snprintf
in the function dtrace_printf_format().  Correct these issues.

Orabug: 30404549
Signed-off-by: David  Mc Lean <david.mclean@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>