Handle @-type and ... for function arglists.
For functions that use the @-type (any type) or varargs, the code
generator was trying to generate type cast code.  Since those two
special cases of type specifications in prototypes are encoded using
invalid types, a SEGV in libctf was reported because we were passing
in invalid arguments.  We now gracefully handle these situations.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>