Store the output buffer offset in the PCB.
As trace output generating functions are encountered in the D script,
we calculate the offset at which trace output should be written into
the output buffer as a running total of all output being generated in
the clause.  This is stored in pcn->pcb_bufoff to be used by the trace
action code generator functions to emit store instructions with
absolute offsets.

Note that string values that may be dynamic (e.g. global variables)
will set aside the maximum string length as buffer space that gets
used because we won't know how much space we need until runtime.  This
is an area that could perhaps be optimized with the output buffer
implemented as a map value.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>