Implement function prologue and epilogue
All BPF functions that implement compiled D clauses need to do some
basic generic work before any D actions can execute, and after all
the D actions have executed.  This is implemented by means of a
generic prologue and epilogue.

The prologue saves the passed in BPF context (ctx, in %r1) and DTrace
context (dctx, in %r2) to the stack.  It also determines the current
CPU id and stores that to the stack.

The epilogue checks whether a fault occured during the D clause
execution, and if so, it discards any recorded trace data for the
probe.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>