Add support for probe BPF trampoline generator.
Each probe provider can (and should) provide a BPF trampoline
generator.  This function is called with an IR list descriptor and
an EPID.  It should append a sequence of BPF instructions to the IR
list, implementing a function that will be called when the probe
fires.  The program needs to set up a DTrace BPF context and then
call the actual implementing function for the compiled D clause.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>