Generate clause predicates as functions
The clause predicates can be compiled as functions that return 0 or 1.
This function can be called from the probe BPF trampoline, and if the
return value is 0, the trampoline can return immediately (the clause
itself is never executed).  This is equivalent to the legacy behaviour
in DTrace.

The introduction of predicates as BPF functions affects the BPF probe
trampoline generator as well because that is where the predicate
function will be called from.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>