Generate trampoline, predicate, and program function together
As a first stage, it makes sense to generate the trampoline, predicate
(if any), and program code as a single instruction block.  BPF expects
programs (incl. any functions that it uses) to be loaded as a single
set of instructions anyway.

As this code matures, it is anticipated that we will migrate to a
design where the trampoline, predicate and program are compiled as
separate functions.  They will then be assembled into a single block of
instructions to be loaded into the kernel through a kind of linking
process.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>