Add support for BPF identifiers, functions, and maps
The linking of final BPF programs requires support for references to
BPF identifier, functions, and maps for use in relocation records.  As
a side effect, this patch also fixes the missing support for relocation
records for kernel and userspace symbols.

We are changing how BPF built-in functions are handled because they can
make use of the more generic identifier support added with this patch.

The disassembler is updated also to account for various changes with the
introduction of relocations.

As a side effect, this patch also changed the name of members in the
dtrace_syminfo_t structure.  This name change was carried during a quite
significant change to that structure (where I renamed the members to be
more in line with newly written code).  The change to the structure was
in the end dropped because an alternative solution was found but the
name change of the existing members remained.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>