Record reclen in DIFO and fix clause return type
We record the trace data record length in the DIFO because we will be
needing it when creating the BPF map for scratch space.  Trace data
is constructed in the scratch space so it needs to be large enough to
hold a full trace record.

This patch also fixes the clause return type (and type size).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>