Fix mem BPF map lookup in prologue
The BPF map lookup in the prologue is supposed to retrieve the value
that corresponds with the current CPU.  Since the map will be created
as a per-CPU map, there is no need to perform the lookup based on the
CPU id as key.  Instead, lookup with key 0 will give us the expected
behaviour, i.e. each CPU will see its own version of the value that is
associated with key 0.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>