Fix relocation bug for kernel symbols used in lddw.
The introduction of BPF relocations and the reworking of support for
relocations caused a problem with the lddw instruction when the
address being loaded from is a kernel symbol (and therefore subject
to relocation) and symbol resolution against the kernel is done in
a static manner.  The code was flagging this as an unexpected asm
relocation for opcode 0x18.  This has been corrected (no error should
be flagged because the relocation is going to be dropped).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>