Ensure FBT probes are associated with modules and weed out duplicates
With the availability of kallsyms with module associations (even for
builtin modules), we can now perform a lookup for FBT probe functions
to ensure we can create the probes with the correct module name.

Given that the source file for FBT probes can (and does) contain
duplicate function names (and no module names to differentiate them),
we need to perform a probe lookup before we create probes to ensure
we do not create duplicate probes.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>