Create the string table as a global resource.
D programs used to be compiled into DIF code where each DIF object
had its own string table.  In the interest of avoiding duplication,
and because we need to do all program processing in userspace anyway
it makes more sense to compile a single string table that is shared
between all compiled programs in a single tracing session.

The string table will be made available to compiled programs (BPF
code) as a strtab BPF map with a singleton (key 0) element that
contains the entire collection of strings concatenated together in
a single chunk of memory.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>