Fix compilation on aarch64
The dt_bpf_context contains a struct that holds the CPU registers,
which is typically 'struct pt_regs'.  However, on aarch64 that struct
is not exposed and we need to use 'struct user_pt_regs' instead.  This
struct is also passed as the BPF context (or part of the BPF context)
for some probe types.

The structure holding the CPU registers is now declared as dt_pt_regs
for all architectures that are supported, hiding the implementation
specific struct.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>